Move manual job code out of app.rb
Based on feedback, move 'manual job' state management out of app.rb into
its own class. Also fix README --profile typo.
