+++ app.rb
@@ -2,6 +2,7 @@ require 'bundler/setup'
 require 'nypl_ruby_util'
 
 require_relative 'lib/state_manager'
+require_relative 'lib/manual_job_state_manager'
 require_relative 'lib/sierra_manager'
 
 def init
@@ -22,7 +23,7 @@ def handle_event(event:, context:)
     # If processing a manual job, create a job-specific state manager:
     if event['manual_job']
         $logger.info "Processing manual job: #{event.to_json}"
-        state = create_manual_job_state_manager event
+        state = ManualJobStateManager.new event
     else
         # Fetch current state from S3
         $logger.info "Loading State from s3"
@@ -39,26 +40,3 @@ def handle_event(event:, context:)
     $logger.info "Processing Complete"
 end
 
-##
-# Create a state manager to handle a single one-off update job
-def create_manual_job_state_manager(event)
-    Class.new do
-        def initialize(event)
-            @event = event
-        end
-
-        def start_time
-            @event['start_time']
-        end
-
-        def end_time
-            @event['end_time']
-        end
-
-        def start_offset
-            @event['start_offset']
-        end
-
-        def set_current_state(execution_time, execution_offset); end
-    end.new event
-end
